% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2.R
\name{add2}
\alias{add2}
\title{Temporary function for package set up purposes only}
\usage{
add2(x)
}
\arguments{
\item{x}{A number}
}
\description{
Temporary function for package set up purposes only
}
\examples{
add2(5)
}
