% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_grid.R
\name{format_grid}
\alias{format_grid}
\title{Format prediction grid using information from a model and the data to which the model was fitted to}
\usage{
format_grid(grid, model, dat.model)
}
\arguments{
\item{grid}{A data frame with columns for spatial coordinates and depth. Each pair of coordinates only occurs once.}

\item{model}{A saved model}

\item{dat.model}{The data that the model was fitted to}
}
\value{
Returns a data frame that can be used to make model predictions, e.g., predict(model, newdata=format_grid(grid,model,dat.model))
}
\description{
Adds parameters to a prediction grid if they are part of a model's formula.
Added parameters include year (number), fyear (factor), month (number), month (factor), depth_scaled (number), and cfvn (factor).
For year and month, the grid is repeated length(unique(x))
}
