% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslSDM-package.R
\docType{package}
\name{gslSDM-package}
\alias{gslSDM}
\alias{gslSDM-package}
\title{gslSDM: R functions to help use package sdmTMB with Gulf of St. Lawrence fisheries data}
\description{
R functions to help use package sdmTMB with Gulf of St. Lawrence fisheries data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoleneSutton/gslSDM}
  \item \url{https://jolenesutton.github.io/gslSDM/}
  \item Report bugs at \url{https://github.com/JoleneSutton/gslSDM/issues}
}

}
\author{
\strong{Maintainer}: Jolene Sutton \email{jolene.sutton@gmail.com}

}
\keyword{internal}
