% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barrier_mesh.R
\name{make_barrier_mesh}
\alias{make_barrier_mesh}
\title{Make an sdmTMB barrier mesh in the Gulf Region}
\usage{
make_barrier_mesh(x, mesh.cutoff)
}
\arguments{
\item{x}{A data frame with a longitude and latitude (epsg:4269).}

\item{mesh.cutoff}{The minimum allowed triangle edge length (see sdmTMB documentation).}
}
\description{
Makes a barrier mesh by applying \code{sdmTMB::make_mesh} and \code{sdmTMBextra::add_barrier_mesh}
with the \code{coastline} shapefile from \code{sglSpatial}. The function \code{add_barrier_mesh}is applied with the default range_fraction = 0.2.
User supplies a data frame with latitude and longitude (epsg:4269).
User supplies a cutoff value to indicate the minimum allowed triangle edge length (see sdmTMB documentation).
The function projects to coordinate reference system: '+proj=utm +zone=20 +datum=NAD83 +units=km +no_defs'.
}
