% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics_sdmTMB.R
\name{plot_diagnostics_sdmTMB}
\alias{plot_diagnostics_sdmTMB}
\title{Generate and plot randomized quantile residuals from sdmTMB models}
\usage{
plot_diagnostics_sdmTMB(model, structure)
}
\arguments{
\item{model}{An sdmTMB model. Assumes model was built from a data set in which spatial coordinates were named 'X' and 'Y'}

\item{structure}{Whether the model as a one-part ("one") structure or a two-part ("hurdle") structure. TARGET FOR FUTURE UPDATES}
}
\value{
Returns a ggplot object or list of ggplot objects
}
\description{
Combines three plots, 1) residuals plotted by X and Y coordinates, 2) histogram of residuals, and 3) a normalized Q-Q plot.
Results are saved to the working directory. CAUTION: See sdmTMB documentation about limitations of using randomized quantile residuals!!
}
\keyword{sdmTMB}
